packages/ethics-gate/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethics-gate"
version = "0.1.0"
description = "BC‑REP Ethics Gate: scores proposed actions with ethics energy & uncertainty, returns permit/deny with rationale."
authors = [{name = "Steven Lanier"}, {name="SEAL Division"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
dependencies = ["pydantic>=2.6","pyyaml>=6.0"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.scripts]
ethics = "ethics_gate.cli:main"
